+++ src/main/java/net/mtrop/doom/struct/io/IOUtils.java
@@ -10,6 +10,9 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.io.RandomAccessFile;
+
+import net.mtrop.doom.struct.DataList;
 
 /**
  * I/O utils.
@@ -17,8 +20,8 @@ import java.io.OutputStream;
  */
 public final class IOUtils
 {
-	/** The relay buffer size, used by relay(). */
-	private static int RELAY_BUFFER_SIZE = 16384;
+	/** The relay buffer used by relay(). */
+	private static final ThreadLocal<byte[]> RELAY_BUFFER = ThreadLocal.withInitial(()->new byte[8192]);
 
 	private IOUtils() {}
 
@@ -50,7 +53,7 @@ public final class IOUtils
 	 */
 	public static int relay(InputStream in, OutputStream out) throws IOException
 	{
-		return relay(in, out, RELAY_BUFFER_SIZE, -1);
+		return relay(in, out, -1);
 	}
 
 	/**
@@ -63,21 +66,82 @@ public final class IOUtils
 	 * This method is thread-safe.
 	 * @param in the input stream to grab data from.
 	 * @param out the output stream to write the data to.
-	 * @param bufferSize the buffer size for the I/O. Must be &gt; 0.
 	 * @param maxLength the maximum amount of bytes to relay, or a value &lt; 0 for no max.
 	 * @return the total amount of bytes relayed.
 	 * @throws IOException if a read or write error occurs.
 	 */
-	public static int relay(InputStream in, OutputStream out, int bufferSize, int maxLength) throws IOException
+	public static int relay(InputStream in, OutputStream out, int maxLength) throws IOException
 	{
 		int total = 0;
 		int buf = 0;
 			
-		byte[] RELAY_BUFFER = new byte[bufferSize];
+		byte[] BUFFER = RELAY_BUFFER.get();
 		
-		while ((buf = in.read(RELAY_BUFFER, 0, Math.min(maxLength < 0 ? Integer.MAX_VALUE : maxLength, bufferSize))) > 0)
+		while ((buf = in.read(BUFFER, 0, Math.min(maxLength < 0 ? Integer.MAX_VALUE : maxLength, BUFFER.length))) > 0)
 		{
-			out.write(RELAY_BUFFER, 0, buf);
+			out.write(BUFFER, 0, buf);
+			total += buf;
+			if (maxLength >= 0)
+				maxLength -= buf;
+		}
+		return total;
+	}
+
+	/**
+	 * Reads from an input stream, reading in a consistent set of data
+	 * and writing it to an open file. The read/write is buffered
+	 * so that it does not bog down the OS's other I/O requests.
+	 * This method finishes when the end of the source stream is reached.
+	 * Note that this may block if the input stream is a type of stream
+	 * that will block if the input stream blocks for additional input.
+	 * This method is thread-safe.
+	 * @param in the input stream to grab data from.
+	 * @param out the file to write the data to.
+	 * @param maxLength the maximum amount of bytes to relay, or a value &lt; 0 for no max.
+	 * @return the total amount of bytes relayed.
+	 * @throws IOException if a read or write error occurs.
+	 */
+	public static int relay(InputStream in, RandomAccessFile out, int maxLength) throws IOException
+	{
+		int total = 0;
+		int buf = 0;
+			
+		byte[] BUFFER = RELAY_BUFFER.get();
+		
+		while ((buf = in.read(BUFFER, 0, Math.min(maxLength < 0 ? Integer.MAX_VALUE : maxLength, BUFFER.length))) > 0)
+		{
+			out.write(BUFFER, 0, buf);
+			total += buf;
+			if (maxLength >= 0)
+				maxLength -= buf;
+		}
+		return total;
+	}
+
+	/**
+	 * Reads from an input stream, reading in a consistent set of data
+	 * and writing it to a {@link DataList}. The read/write is buffered
+	 * so that it does not bog down the OS's other I/O requests.
+	 * This method finishes when the end of the source stream is reached.
+	 * Note that this may block if the input stream is a type of stream
+	 * that will block if the input stream blocks for additional input.
+	 * This method is thread-safe.
+	 * @param in the input stream to grab data from.
+	 * @param out the file to write the data to.
+	 * @param maxLength the maximum amount of bytes to relay, or a value &lt; 0 for no max.
+	 * @return the total amount of bytes relayed.
+	 * @throws IOException if a read or write error occurs.
+	 */
+	public static int relay(InputStream in, DataList out, int maxLength) throws IOException
+	{
+		int total = 0;
+		int buf = 0;
+			
+		byte[] BUFFER = RELAY_BUFFER.get();
+		
+		while ((buf = in.read(BUFFER, 0, Math.min(maxLength < 0 ? Integer.MAX_VALUE : maxLength, BUFFER.length))) > 0)
+		{
+			out.append(BUFFER, 0, buf);
 			total += buf;
 			if (maxLength >= 0)
 				maxLength -= buf;